Large Storage refactoring and new 'todo' plugin.
Storage was refactored to use custom Shelve and pickle objects.
Now it is safe to have references to the Bot object in yours
requests and to save them into the storage.

Also, each request class now have to have `get_user` method.

Todo plugin was added. It behaves as a reminder.